dom/tests/mochitest/chrome/file_DOM_element_instanceof.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<window title="Mozilla Bug 824917"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <iframe type="content"></iframe>

  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 799299 **/
  var SimpleTest = window.arguments[0].SimpleTest;
  var ok = window.arguments[0].ok;

  // Note: We can't use frames[0] here because the type="content" attribute
  // isolates it into a separate browsing context hierarchy.
  var doc = document.querySelector("iframe").contentDocument;
  ok(HTMLBodyElement.isInstance(doc.createElement("body")),
     "Should be instance of HTMLBodyElement");
  ok(HTMLDivElement.isInstance(doc.createElement("div")),
     "Should be instance of HTMLDivElement");
  ok(HTMLFrameSetElement.isInstance(doc.createElement("frameset")),
     "Should be instance of HTMLFrameSetElement");
  ok(HTMLHeadingElement.isInstance(doc.createElement("h1")),
     "Should be instance of HTMLHeadingElement");
  ok(HTMLLabelElement.isInstance(doc.createElement("label")),
     "Should be instance of HTMLLabelElement");

  window.close();
  window.arguments[0].SimpleTest.finish();
  ]]>
  </script>
</window>
